requirements.txt
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
aiopg==1.0.0
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
motor==2.1.0
multidict==4.7.6
psycopg2-binary==2.8.5
pymongo==3.10.1
PyYAML==5.3.1
SQLAlchemy==1.3.18
yarl==1.4.2
